feat: add AbstractService & AbstractServiceOptions to @directus/types
