+++ api/src/services/graphql/index.ts
@@ -1,10 +1,10 @@
 import { useEnv } from '@directus/env';
-import type { Accountability, Item, Query, SchemaOverview } from '@directus/types';
+import type { AbstractServiceOptions, Accountability, Item, Query, SchemaOverview } from '@directus/types';
 import type { ExecutionResult, FormattedExecutionResult, GraphQLSchema } from 'graphql';
 import { NoSchemaIntrospectionCustomRule, execute, specifiedRules, validate } from 'graphql';
 import type { Knex } from 'knex';
 import getDatabase from '../../database/index.js';
-import type { AbstractServiceOptions, GraphQLParams } from '../../types/index.js';
+import type { GraphQLParams } from '../../types/index.js';
 import { getService } from '../../utils/get-service.js';
 import { formatError } from './errors/format.js';
 import { GraphQLExecutionError, GraphQLValidationError } from './errors/index.js';
