+++ api/src/services/mail/index.ts
@@ -1,6 +1,6 @@
 import { useEnv } from '@directus/env';
 import { InvalidPayloadError } from '@directus/errors';
-import type { Accountability, SchemaOverview } from '@directus/types';
+import type { AbstractServiceOptions, Accountability, SchemaOverview } from '@directus/types';
 import fse from 'fs-extra';
 import type { Knex } from 'knex';
 import { Liquid } from 'liquidjs';
@@ -10,7 +10,6 @@ import { fileURLToPath } from 'url';
 import getDatabase from '../../database/index.js';
 import { useLogger } from '../../logger/index.js';
 import getMailer from '../../mailer.js';
-import type { AbstractServiceOptions } from '../../types/index.js';
 import { Url } from '../../utils/url.js';
 import emitter from '../../emitter.js';
