+++ api/src/services/notifications.ts
@@ -1,9 +1,9 @@
 import { useEnv } from '@directus/env';
-import type { Notification, PrimaryKey } from '@directus/types';
+import type { AbstractServiceOptions, Notification, PrimaryKey } from '@directus/types';
 import { useLogger } from '../logger/index.js';
 import { fetchRolesTree } from '../permissions/lib/fetch-roles-tree.js';
 import { fetchGlobalAccess } from '../permissions/modules/fetch-global-access/fetch-global-access.js';
-import type { AbstractServiceOptions, MutationOptions } from '../types/index.js';
+import type { MutationOptions } from '../types/index.js';
 import { md } from '../utils/md.js';
 import { Url } from '../utils/url.js';
 import { ItemsService } from './items.js';
