+++ api/src/services/operations.ts
@@ -1,6 +1,6 @@
-import type { Item, OperationRaw, PrimaryKey } from '@directus/types';
+import type { AbstractServiceOptions, Item, OperationRaw, PrimaryKey } from '@directus/types';
 import { getFlowManager } from '../flows.js';
-import type { AbstractServiceOptions, MutationOptions } from '../types/index.js';
+import type { MutationOptions } from '../types/index.js';
 import { ItemsService } from './items.js';
 
 export class OperationsService extends ItemsService<OperationRaw> {
