+++ api/src/services/permissions.ts
@@ -1,5 +1,5 @@
 import { ForbiddenError } from '@directus/errors';
-import type { Item, ItemPermissions, Permission, PrimaryKey, Query } from '@directus/types';
+import type { AbstractServiceOptions, Item, ItemPermissions, Permission, PrimaryKey, Query } from '@directus/types';
 import { uniq } from 'lodash-es';
 import { clearSystemCache } from '../cache.js';
 import { fetchPermissions } from '../permissions/lib/fetch-permissions.js';
@@ -7,7 +7,7 @@ import { fetchPolicies } from '../permissions/lib/fetch-policies.js';
 import { withAppMinimalPermissions } from '../permissions/lib/with-app-minimal-permissions.js';
 import type { ValidateAccessOptions } from '../permissions/modules/validate-access/validate-access.js';
 import { validateAccess } from '../permissions/modules/validate-access/validate-access.js';
-import type { AbstractServiceOptions, MutationOptions } from '../types/index.js';
+import type { MutationOptions } from '../types/index.js';
 import type { QueryOptions } from './items.js';
 import { ItemsService } from './items.js';
