+++ api/src/services/revisions.ts
@@ -1,6 +1,6 @@
 import { ForbiddenError, InvalidPayloadError } from '@directus/errors';
-import type { Item, PrimaryKey } from '@directus/types';
-import type { AbstractServiceOptions, MutationOptions } from '../types/index.js';
+import type { AbstractServiceOptions, Item, PrimaryKey } from '@directus/types';
+import type { MutationOptions } from '../types/index.js';
 import { ItemsService } from './items.js';
 
 export class RevisionsService extends ItemsService {
