+++ api/src/services/specifications.ts
@@ -2,7 +2,14 @@ import { useEnv } from '@directus/env';
 import formatTitle from '@directus/format-title';
 import { spec } from '@directus/specs';
 import { isSystemCollection } from '@directus/system-data';
-import type { Accountability, FieldOverview, Permission, SchemaOverview, Type } from '@directus/types';
+import type {
+	AbstractServiceOptions,
+	Accountability,
+	FieldOverview,
+	Permission,
+	SchemaOverview,
+	Type,
+} from '@directus/types';
 import { getRelation } from '@directus/utils';
 import { version } from 'directus/version';
 import type { Knex } from 'knex';
@@ -20,7 +27,6 @@ import getDatabase from '../database/index.js';
 import { fetchPermissions } from '../permissions/lib/fetch-permissions.js';
 import { fetchPolicies } from '../permissions/lib/fetch-policies.js';
 import { fetchAllowedFieldMap } from '../permissions/modules/fetch-allowed-field-map/fetch-allowed-field-map.js';
-import type { AbstractServiceOptions } from '../types/index.js';
 import { getRelationType } from '../utils/get-relation-type.js';
 import { reduceSchema } from '../utils/reduce-schema.js';
 import { GraphQLService } from './graphql/index.js';
