+++ api/src/services/translations.ts
@@ -1,8 +1,7 @@
-import type { Item, PrimaryKey } from '@directus/types';
+import type { AbstractServiceOptions, Item, PrimaryKey } from '@directus/types';
 import getDatabase from '../database/index.js';
 import { InvalidPayloadError } from '@directus/errors';
 import type { MutationOptions } from '../types/items.js';
-import type { AbstractServiceOptions } from '../types/services.js';
 import { ItemsService } from './items.js';
 
 export class TranslationsService extends ItemsService {
