+++ api/src/types/index.ts
@@ -9,6 +9,5 @@ export * from './items.js';
 export * from './meta.js';
 export * from './migration.js';
 export * from './revision.js';
-export * from './services.js';
 export * from './snapshot.js';
 export * from './webhooks.js';
