--- api/src/types/services.ts
@@ -1,28 +1,0 @@
-import type { Accountability, Item, PrimaryKey, Query, SchemaOverview } from '@directus/types';
-import type { Knex } from 'knex';
-
-export type AbstractServiceOptions = {
-	knex?: Knex | undefined;
-	accountability?: Accountability | null | undefined;
-	schema: SchemaOverview;
-	nested?: string[];
-};
-
-export interface AbstractService {
-	knex: Knex;
-	accountability: Accountability | null | undefined;
-	nested: string[];
-
-	createOne(data: Partial<Item>): Promise<PrimaryKey>;
-	createMany(data: Partial<Item>[]): Promise<PrimaryKey[]>;
-
-	readOne(key: PrimaryKey, query?: Query): Promise<Item>;
-	readMany(keys: PrimaryKey[], query?: Query): Promise<Item[]>;
-	readByQuery(query: Query): Promise<Item[]>;
-
-	updateOne(key: PrimaryKey, data: Partial<Item>): Promise<PrimaryKey>;
-	updateMany(keys: PrimaryKey[], data: Partial<Item>): Promise<PrimaryKey[]>;
-
-	deleteOne(key: PrimaryKey): Promise<PrimaryKey>;
-	deleteMany(keys: PrimaryKey[]): Promise<PrimaryKey[]>;
-}
